Added pyfunction eval and refactored
Conflicts:
	symengine/lib/symengine_wrapper.pyx
